Grass behaviour (without Sense/Think/Act)
Added basic grass behaviour.
Grass has a health value from 0 to 1 and three states (seed, mature, dirt). At the start, a random number of tiles are set to seed, the rest is dirt. If in seed, the grass gains health until it reaches mature. When mature, the grass slowly loses health until it dies and becomes dirt. During this process it will try to spread to neighbouring cells.