Added Observer pattern and Test Project
Added Observer, Subject and Event to implement the Observer pattern similarly to how it is described in the assignment description.
Uses std::list for the observer list inside Subject. Since this is a new class for me, I tried to test it using Visual Studio cpp tests but it does not work yet.